shadys-nails-backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Environment & Settings
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2  # Pin to 3.2.2 for passlib compatibility

# HTTP Client
requests>=2.31.0

# Google OAuth
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
